backend/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.0
motor==3.7.0
python-dotenv==1.1.0
pydantic-settings==2.9.1
openai==1.76.0
beautifulsoup4==4.12.3
requests==2.32.0
pymongo==4.12.0
langchain==0.3.24 # Add langchain core
langchain-openai==0.3.14 # Add langchain openai integration
chromadb==1.0.7 # Add chromadb vector store
tiktoken==0.9.0 # Add tiktoken tokenizer
langchain-community==0.3.22